requirements.txt
Faker==25.8.0
pytest==8.2.1
python-dotenv==1.0.1
selenium==4.21.0
webdriver_manager==4.0.1
pytest-xdist==3.6.1
allure-pytest
